Major revisions - move to version 2.
This now has a clearer layout, with all 1000+ lines of
code for functions being moved to aerosol_size_distributions.py.
This will hopefully improve readibility. Minor quality of life
improvements and updates to information and the way calibrations are
generate in accordance with Graham's new upload policy.
Example FAAM calibrations are now included.